not crawling friends if already present.

+++ src/fastCrawler/crawl_friends.py
@@ -4,6 +4,7 @@ from twython import Twython, TwythonRateLimitError
 #import config
 import json
 import thread
+import os
 
 config_file = open('config.json','r')
 configs = json.load(config_file)
@@ -25,19 +26,26 @@ def start_crawler(app_key, app_secret, in_file):
     i=0
     while True and i<len(all_users):
         user = all_users[i]
-        try:
-            friends = twitter.get_friends_ids(user_id=user,count=5000)
-            #print friends
-            out = open('../../data/friend_id/'+str(user),'w')
-            json.dump(friends,out)
-            i+=1
-            log (user,i)
-        except:
-            reset = int(twitter.get_lastfunction_header('x-rate-limit-reset'))
-            msg =  "waiting for "+str(reset - time.time())+ ' sec'
-            log(user,msg)
-            wait = max(reset - time.time(), 0) + 10 # addding 10 second pad
-            time.sleep(wait)
+        file_name = '../../data/friend_id/'+str(user)
+        if os.path.isfile(file_name):
+            log(user,str(i) + " data already present")
+        else:
+            try:
+                # Crawl user friends
+                out = open(file_name,'w')
+                friends = twitter.get_friends_ids(user_id=user,count=5000)
+                json.dump(friends,out)
+                log (user,i)
+            except:
+                # rate limit reached
+                 reset = int(twitter.get_lastfunction_header('x-rate-limit-reset'))
+                 msg =  "waiting for "+str(reset - time.time())+ ' sec'
+                 log(user,msg)
+                 wait = max(reset - time.time(), 0) + 10 # addding 10 second pad
+                 time.sleep(wait)
+                 i-=1
+        i+=1
+
 
 print "Crawler initialized:" ,sys.argv[1], sys.argv[2]
 start_crawler(sys.argv[1], sys.argv[2],sys.argv[3])
